requirements.txt
flask==2.3.0
PyJWT==2.10.1
flasgger==0.9.7.1
pytest==7.4.0
pytest-cov==4.1.0
gunicorn==21.0.0
